+++ src/DepthOnlyPass.h
@@ -24,7 +24,7 @@ public:
 
 	static ComPtr<ID3D12RootSignature> BuildRootSignature( ID3D12Device& device );
 
-	static void BuildData( DXGI_FORMAT dsv_format, float bias, bool back_culling, ID3D12Device& device, ComPtr<ID3D12PipelineState>& pso, ComPtr<ID3D12RootSignature>& rootsig );
+	static void BuildData( DXGI_FORMAT dsv_format, int bias, bool back_culling, ID3D12Device& device, ComPtr<ID3D12PipelineState>& pso, ComPtr<ID3D12RootSignature>& rootsig );
 
 	static ComPtr<ID3DBlob> LoadAndCompileVertexShader();
